pyproject.toml
[project]
name = "aegis"
version = "0.1.0"
description = "AEGIS: Agentic Exploit Generation in Isolated Sandboxes"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.86.0",
    "pydantic>=2.12.5",
]

[project.scripts]
aegis = "aegis.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aegis"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pylint>=3.0.0",
    "pyyaml>=6.0",
]

[tool.pylint.main]
source-roots = ["src"]

[tool.pylint.messages_control]
disable = [
    "fixme",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "line-too-long",
    "logging-fstring-interpolation",
    "too-few-public-methods",
    "broad-exception-caught",
    "protected-access",
    "redefined-outer-name",
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
testpaths = ["tests"]
addopts = "-m 'not benchmark'"
markers = [
    "benchmark: ground truth evaluation tests (deselect with '-m not benchmark')",
]
